+++ controllable/ControllableContainer.h
@@ -77,7 +77,7 @@ public:
 	UndoableAction * addUndoableControllable(Controllable * c, bool onlyReturnAction = false);
 	Controllable* addControllable(Controllable * c, int index = -1);
 	Parameter* addParameter(Parameter * p, int index = -1);
-	FloatParameter * addFloatParameter(const String &niceName, const String &description, const float &initialValue, const float &minValue = INT32_MIN, const float &maxValue = INT32_MAX, const bool &enabled = true);
+	FloatParameter * addFloatParameter(const String &niceName, const String &description, const double &initialValue, const double &minValue = INT32_MIN, const double &maxValue = INT32_MAX, const bool &enabled = true);
 	IntParameter * addIntParameter(const String &niceName, const String &description, const int &initialValue, const int &minValue = INT32_MIN, const int &maxValue = INT32_MAX, const bool &enabled = true);
 	BoolParameter * addBoolParameter(const String &niceName, const String &description, const bool &value, const bool &enabled = true);
 	StringParameter * addStringParameter(const String &niceName, const String &description, const String &value, const bool &enabled = true);
